Move program generator to its own package
